Translate UI to Russian and add registration
